+++ src/state/entity.ts
@@ -37,6 +37,7 @@ import './entity/ig_ENTITY_ItemDestruct'
 import './entity/ig_ENTITY_XenoDialog'
 import './entity/sc_PartyMemberEntity'
 import './entity/ig_ENTITY_Prop'
+import './entity/sc_FoodIconEntity'
 
 type EntityStateUnion = EntityStates[keyof EntityStates]